# Week-4/3.WebApi_Handson/SwaggerDemoApi/Models/Employee.cs
using System;
using System.Collections.Generic;

namespace SwaggerDemoApi.Models
{
    public class Employee
    {
        public int Id { get; set; }
        public string Name { get; set; }
        public int Salary { get; set; }
        public bool Permanent { get; set; }
        public Department Department { get; set; }
        public List<Skill> Skills { get; set; }
        public DateTime DateOfBirth { get; set; }
    }
}